requirements.txt
pytorch-lightning
torchvision
wandb
